+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Properly convert elixir nil to erlang :null in query params
+
 ## [0.4.1] - 2021-09-02
 
 ### Added